Configure Undertow to eagerly initialize Filters by default
Closes gh-9232
